// front/admin/project/src/open_api/models/Entry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI Tutorial
 * OpenAPI Tutorial by halhorn
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 応募
 * @export
 * @interface Entry
 */
export interface Entry {
    /**
     * id
     * @type {number}
     * @memberof Entry
     */
    id?: number;
    /**
     * 会員id
     * @type {number}
     * @memberof Entry
     */
    userId?: number;
    /**
     * 仕事id
     * @type {number}
     * @memberof Entry
     */
    jobId?: number;
}

export function EntryFromJSON(json: any): Entry {
    return EntryFromJSONTyped(json, false);
}

export function EntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): Entry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'userId': !exists(json, 'userId') ? undefined : json['userId'],
        'jobId': !exists(json, 'jobId') ? undefined : json['jobId'],
    };
}

export function EntryToJSON(value?: Entry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'userId': value.userId,
        'jobId': value.jobId,
    };
}